+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/indexes/SearchIndexerAsyncClient.java
@@ -17,9 +17,11 @@ import com.azure.search.documents.implementation.converters.SearchIndexerConvert
 import com.azure.search.documents.implementation.converters.SearchIndexerDataSourceConverter;
 import com.azure.search.documents.implementation.util.MappingUtils;
 import com.azure.search.documents.indexes.implementation.SearchServiceClientImpl;
+import com.azure.search.documents.indexes.implementation.models.DocumentKeysOrIds;
 import com.azure.search.documents.indexes.implementation.models.ListDataSourcesResult;
 import com.azure.search.documents.indexes.implementation.models.ListIndexersResult;
 import com.azure.search.documents.indexes.implementation.models.ListSkillsetsResult;
+import com.azure.search.documents.indexes.implementation.models.SkillNames;
 import com.azure.search.documents.indexes.models.CreateOrUpdateDataSourceConnectionOptions;
 import com.azure.search.documents.indexes.models.CreateOrUpdateIndexerOptions;
 import com.azure.search.documents.indexes.models.CreateOrUpdateSkillsetOptions;
@@ -29,6 +31,7 @@ import com.azure.search.documents.indexes.models.SearchIndexerSkillset;
 import com.azure.search.documents.indexes.models.SearchIndexerStatus;
 import reactor.core.publisher.Mono;
 
+import java.util.List;
 import java.util.function.Function;
 
 import static com.azure.core.util.FluxUtil.monoError;
@@ -1057,6 +1060,60 @@ public class SearchIndexerAsyncClient {
     }
 
     /**
+     * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
+     *
+     * @param indexerName The name of the indexer to reset documents for.
+     * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
+     * payload will be queued to be re-ingested.
+     * @param documentKeys Document keys to be reset.
+     * @param datasourceDocumentIds Datasource document identifiers to be reset.
+     * @return A response signalling completion.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<Void> resetDocuments(String indexerName, boolean overwrite, List<String> documentKeys,
+        List<String> datasourceDocumentIds) {
+        return withContext(context -> resetDocumentsWithResponse(indexerName, overwrite, documentKeys,
+            datasourceDocumentIds, context))
+            .map(Response::getValue);
+    }
+
+    /**
+     * Resets specific documents in the datasource to be selectively re-ingested by the indexer.
+     *
+     * @param indexer The indexer to reset documents for.
+     * @param overwrite If false, keys or IDs will be appended to existing ones. If true, only the keys or IDs in this
+     * payload will be queued to be re-ingested.
+     * @param documentKeys Document keys to be reset.
+     * @param datasourceDocumentIds Datasource document identifiers to be reset.
+     * @return A response signalling completion.
+     * @throws NullPointerException If {@code indexer} is null.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<Response<Void>> resetDocumentsWithResponse(SearchIndexer indexer, boolean overwrite,
+        List<String> documentKeys, List<String> datasourceDocumentIds) {
+        if (indexer == null) {
+            return monoError(logger, new NullPointerException("'indexer' cannot be null."));
+        }
+
+        return withContext(context -> resetDocumentsWithResponse(indexer.getName(), overwrite, documentKeys,
+            datasourceDocumentIds, context));
+    }
+
+    Mono<Response<Void>> resetDocumentsWithResponse(String indexerName, Boolean overwrite, List<String> documentKeys,
+        List<String> datasourceDocumentIds, Context context) {
+        try {
+            DocumentKeysOrIds documentKeysOrIds = new DocumentKeysOrIds()
+                .setDocumentKeys(documentKeys)
+                .setDatasourceDocumentIds(datasourceDocumentIds);
+
+            return restClient.getIndexers()
+                .resetDocsWithResponseAsync(indexerName, overwrite, documentKeysOrIds, null, context);
+        } catch (RuntimeException ex) {
+            return monoError(logger, ex);
+        }
+    }
+
+    /**
      * Creates a new skillset in an Azure Cognitive Search service.
      *
      * <p><strong>Code Sample</strong></p>
@@ -1471,6 +1528,45 @@ public class SearchIndexerAsyncClient {
                 .deleteWithResponseAsync(skillsetName, eTag, null, null, context)
                 .onErrorMap(MappingUtils::exceptionMapper)
                 .map(Function.identity());
+        } catch (RuntimeException ex) {
+            return monoError(logger, ex);
+        }
+    }
+
+    /**
+     * Resets skills in an existing skillset in an Azure Cognitive Search service.
+     *
+     * @param skillsetName The name of the skillset to reset.
+     * @param skillNames The skills to reset.
+     * @return A response signalling completion.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<Void> resetSkills(String skillsetName, List<String> skillNames) {
+        return withContext(context -> resetSkillsWithResponse(skillsetName, skillNames, context)
+            .flatMap(FluxUtil::toMono));
+    }
+
+    /**
+     * Resets skills in an existing skillset in an Azure Cognitive Search service.
+     *
+     * @param skillset The skillset to reset.
+     * @param skillNames The skills to reset.
+     * @return A response signalling completion.
+     * @throws NullPointerException If {@code skillset} is null.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    public Mono<Response<Void>> resetSkillsWithResponse(SearchIndexerSkillset skillset, List<String> skillNames) {
+        if (skillset == null) {
+            return monoError(logger, new NullPointerException("'skillset' cannot be null."));
+        }
+
+        return withContext(context -> resetSkillsWithResponse(skillset.getName(), skillNames, context));
+    }
+
+    Mono<Response<Void>> resetSkillsWithResponse(String skillsetName, List<String> skillNames, Context context) {
+        try {
+            return restClient.getSkillsets()
+                .resetSkillsWithResponseAsync(skillsetName, new SkillNames().setSkillNames(skillNames), null, context);
         } catch (RuntimeException ex) {
             return monoError(logger, ex);
         }